requirements.txt
# SOC Project Dependencies
# Install with: pip install -r requirements.txt

# Core data
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0       # Parquet read/write
pyyaml>=6.0

# ML models (Stage 4 — model training)
scikit-learn>=1.3.0   # Isolation Forest, LOF, preprocessing
torch>=2.1.0          # Autoencoder
# OR: tensorflow>=2.13.0  (alternative to torch for autoencoder)

# Real AWS connection
boto3>=1.34.0
botocore>=1.34.0

# RAG layer (coordinate with your friend's setup)
chromadb>=0.4.0
neo4j>=5.14.0

# Optional: faster Parquet
fastparquet>=2023.10.1

# Dev/testing
pytest>=7.4.0
jupyter>=1.0.0        # For exploratory analysis
